修改目录

+++ changedetection/utils_func/metrics.py
@@ -91,7 +91,7 @@ class Evaluator(object):
 
     def _generate_matrix(self, gt_image, pre_image):
         mask = (gt_image >= 0) & (gt_image < self.num_class)
-        label = self.num_class * gt_image[mask].astype('int64') + pre_image[mask]
+        label = self.num_class * gt_image[mask] + pre_image[mask]
         count = np.bincount(label, minlength=self.num_class ** 2)
         confusion_matrix = count.reshape(self.num_class, self.num_class)
         return confusion_matrix
